+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG
 =========
 
  * added task doctrine:compile to package the library into an unique cached file
+ * fixed task project:optimize
  * added unit tests for yaml and event classes
  * added sfWidgetFormDoctrineArrayChoice widget
  * added sfFormObject::saveObject() method